\name{jointNmix-package}
\alias{jointNmix-package}
\encoding{UTF8}
\docType{package}
\title{
\packageTitle{jointNmix}
}
\description{
\packageDescription{jointNmix}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{jointNmix}
\packageIndices{jointNmix}
}
\author{
\packageAuthor{jointNmix}

Maintainer: \packageMaintainer{jointNmix}
}
\references{Moral, R.A., Hinde, J., \enc{Demétrio}{Demetrio}, C.G.B., Reigada, C. and Godoy, W.A.C. (submitted) Models for jointly estimating abundance of two unmarked site-associated species subject to imperfect detection.}

\keyword{ package }